Bug #4517: Emit observer event to filter the Drag+Drop url list
This patch creates an "on-modify-drag-list" observer that addons can listen
to. For us, it supports Torbutton code that filters out Drag+Drop mime types
that the OS Desktop sniffs and attempts to load without Tor.

Such proxy bypass behavior is immediate on Mac and Ubuntu: you don't even have
to release the object for it to get sniffed and cause the OS to load it
without Tor. In fact, accidentally clicking for too long on an image is enough
to cause proxy bypass on those systems.