Bug #2874: Block Components.interfaces from content
This patch removes the ability of content script to access
Components.interfaces.*. Implemented by removing the Components.interfaces
shim introduced in Parts 4, 6, and 7 of the patch in
https://bugzilla.mozilla.org/show_bug.cgi?id=790732.

Components.interfaces is useful for fingerprinting the platform, OS, and
Firefox version.

It appears to have no other legitimate use. See also:
https://bugzilla.mozilla.org/show_bug.cgi?id=429070

Includes regression test.