Bug 13594: Windows updater depends on msvcr100.dll
On Windows, updater.exe failed to start if a copy of msvcr100.dll was not
installed in the system directory.  We now append to the PATH the directory
that contains our copy of msvcr100.dll; that is, the Browser/ directory that
contains firefox.exe.  That same directory contains a copy of libssp-0.dll,
which updater.exe also depends on; both DLL dependencies are now satisfied
from the Browser/ directory.  Previously, the libssp-0.dll dependency was being
satisifed from the .../Browser/TorBrowser/Tor directory which Tor Launcher
adds to the path (and typically the updater is run from within a browser
session or after a restart during which the PATH is preserved).