Bug 15910: Disable GMPs for now
Until there are GMPs that can get built reproducible AND whose source
code is inspectable we disable the underlying install and update
mechanisms.

We can't use OpenH264 and Adobe's CDM in Tor Browser anyway
as the former is currently only used for WebRTC which we disable for
content and the latter is neither available for Linux nor OS X and does
not run on Windows as we can't compile the sandbox for it yet which is a
hard requirement.