Bug 16528: Prevent indexedDB Modernizr breakage (e10s highrisk).
This change should make the indexedDB failure mode for pref checks equivalent
to the private browsing failure mode. Sites that use Modernizr are accustomed
to the failure modes for Private Browsing usage of IndexedDB, but not for when
the pref is disabled.

This patch may cause serious issues with e10s in the future. We'll need to
keep an eye on it for FF52.