Bug 16236: Windows updater: avoid writing to the registry.
Mozilla moves "in use" files that cannot be deleted during an
update to a "tobedeleted" directory and then makes a call to
MoveFileEx(..., MOVEFILE_DELAY_UNTIL_REBOOT) to request that
the file be deleted later.  To avoid writing to the Windows Registry,
we simply try to remove the "tobedeleted" directory and its contents
during browser startup.