Bug 20589: Adding new MAR signing key
We make the certificate for the secondary key the new primary one,
partly due to #15532 and add the certificate for the new key as the
secondary one.

Also Bug 18008: Create a new MAR Signing key

This is part of our plan to rotate the main key used for signing MAR
files about once a year to make. This is a defense in depth measure as
there is no official way for revoking MAR signing keys.