Bug 18170: After update, only changelog tab shown
When in permanent private browsing mode, always return false
for isAutomaticRestoreEnabled. This ensures that there will
not be any confusion inside nsBrowserContentHandler.defaultArgs
as to whether a one time session restore will occur.

Also, for consistency and in case someone looks at the pref,
avoid setting browser.sessionstore.resume_session = true during
browser shutdown.

This bug occurred when staging was not used during the update
process. On Windows it always occurred because staging is not
used even when it should be (see #18292).