Bug 18912: add automated tests for updater cert pinning
Add a chrome mochitest that verifies that a key pinning error
is generated when the updater is configured to use one of the
"bad pin" servers that are predefined by the test framework.

Note that this test is affected by bug 18087, which means that
security.nocertdb must be set to false before running this test.