fixup! TB4: Tor Browser's Firefox preference overrides.
We avoid fingerprinting risks associated with Reader View while keeping
the feature enabled for users that really want to have it. This is done
by making sure that documents are not parsed on load as this is disabled
on some devices due to memory consumption and we pretend that everybody
has already been made aware of that feature. (bug 18950)