fixup! Bug #4234: Use the Firefox Update Process for Tor Browser.
Bug 22041: libmozsandbox.so error after upgrading to 7.0a3

Fix a problem with unstaged updates on Linux and OSX where the original
files were not restored after a partial update failed to apply due to
changes the user made within their installation. This would leave the
browser in a state where it had a mix of old files and new (patched)
files, which can cause the browser to not be able to start, and
therefore not be able to fallback to a complete update.