fixup! Bug #4234: Use the Firefox Update Process for Tor Browser.
Disable the updater's OSX privilege elevation feature. If the user
does not have write permission to the .app directory, they will be
prompted to download a new copy of the browser. Fixes bug 21940.

As defense-in-depth measures, we also removed the code that is
used by Firefox to launch an OSX process with administrator privileges
as well as the extra copy of the updater executable that was located
under TorBrowser.app/Contents/Library/LaunchServices/.