Bug 25458: UI customization half-broken in Tor Browser 8.0a3
Fixed fallout from #25147.  Some of the functionality of the UI
Customize screen is dependent on setting innerHTML on divs.  Since the
code runs in the System context, the HTML was being sanitized and broke
page functionality.  The offending statements have been switched to the
unsafeSetInnerHTML() method which bypasses sanitization in System
context.