fixup! TB4: Tor Browser's Firefox preference overrides.
In Mozilla Bug 1217290, a new WebGL mode was introduced specifically
for privacy.resistFingerprinting = true, which is supposed to give
most machines the same WebGL profile.

Previously, webgl.min_capability_mode was use for this purpose,
but now that pref is intended strictly for browser testing and it
overrides the privacy.resistFingerprinting setting. Furthermore,
webgl.min_capability_mode can cause webgl to fail unnecessarily.

So we should not be enabling webgl.min_capability_mode any more in
Tor Browser. This patch removes that pref setting so it reverts to
Firefox's default value (false).